Simplified JsStreamingData
